refactor: reorganize session lifecycle code into dedicated package
