+++ src/deephaven_mcp/sessions/_lifecycle/shared.py
