+++ tests/test_sessions.py
@@ -20,7 +20,6 @@ from deephaven_mcp.sessions._sessions import (
     SessionCreationError,
     SessionManager,
     create_session,
-    get_session_parameters,
 )
 
 
@@ -44,18 +43,6 @@ def session_manager(mock_config_manager):
 # --- Additional Robustness Tests ---
 
 
-@pytest.mark.asyncio
-async def test_get_session_parameters_missing_fields():
-    # Should not raise if only host provided
-    params = await get_session_parameters({"host": "localhost"})
-    assert params["host"] == "localhost"
-    # Should fill defaults
-    assert params["auth_type"] == "Anonymous"
-    assert params["auth_token"] == ""
-    assert params["never_timeout"] is False
-    assert params["session_type"] == "python"
-    assert params["use_tls"] is False
-
 
 @pytest.mark.asyncio
 async def test_create_session_error_handling():
@@ -67,7 +54,7 @@ async def test_create_session_error_handling():
         with pytest.raises(SessionCreationError) as exc_info:
             await create_session(host="localhost")
         # Check error message contains context and original error
-        assert "Failed to create Deephaven Session" in str(exc_info.value)
+        assert "Failed to create Deephaven Community (Core) Session" in str(exc_info.value)
         assert "fail" in str(exc_info.value) or (exc_info.value.__cause__ and "fail" in str(exc_info.value.__cause__))
 
 
@@ -98,7 +85,7 @@ async def test_session_manager_concurrent_get_or_create_session():
 
     # Success case: patch Session in _lifecycle_community to return mock_session
     with patch(
-        "deephaven_mcp.sessions._lifecycle_community.Session",
+        "deephaven_mcp.sessions._lifecycle.community.Session",
         new=MagicMock(return_value=mock_session),
     ):
         # Fire off multiple concurrent requests for the same worker
@@ -118,12 +105,12 @@ async def test_session_manager_concurrent_get_or_create_session_failure():
     )
     mgr = SessionManager(mock_cfg_mgr)
     with patch(
-        "deephaven_mcp.sessions._lifecycle_community.Session",
+        "deephaven_mcp.sessions._lifecycle.community.Session",
         new=MagicMock(side_effect=RuntimeError("fail")),
     ):
         with pytest.raises(SessionCreationError) as exc_info:
             await mgr.get_or_create_session("workerX")
-        assert "Failed to create Deephaven Session" in str(exc_info.value)
+        assert "Failed to create Deephaven Community (Core) Session" in str(exc_info.value)
         assert "fail" in str(exc_info.value) or (exc_info.value.__cause__ and "fail" in str(exc_info.value.__cause__))
 
 
@@ -167,117 +154,19 @@ async def test_clear_all_sessions_calls_close(session_manager):
     session.close.assert_called_once()
 
 
-@pytest.mark.asyncio
-async def test_get_session_parameters_with_and_without_files():
-    mgr = SessionManager(mock_config_manager)
-    from unittest.mock import AsyncMock, patch
-
-    with patch(
-        "deephaven_mcp.sessions._lifecycle_community.load_bytes",
-        new=AsyncMock(return_value=b"binary"),
-    ):
-        # All fields present (as file paths)
-        cfg = {
-            "host": "localhost",
-            "port": 10000,
-            "auth_type": "token",
-            "auth_token": "tok",
-            "never_timeout": True,
-            "session_type": "python",
-            "use_tls": True,
-            "tls_root_certs": "/tmp/root.pem",
-            "client_cert_chain": "/tmp/chain.pem",
-            "client_private_key": "/tmp/key.pem",
-        }
-        params = await get_session_parameters(cfg)
-        assert params["tls_root_certs"] == b"binary"
-        assert params["client_cert_chain"] == b"binary"
-        assert params["client_private_key"] == b"binary"
-        # No files present
-        cfg = {"host": "localhost"}
-        params = await get_session_parameters(cfg)
-        assert params["host"] == "localhost"
-
-
-@pytest.mark.asyncio
-async def test_get_session_parameters_file_error():
-    # Patch load_bytes to raise
-    async def raise_io(path):
-        raise IOError("fail")
-
-    with patch("deephaven_mcp.sessions._sessions.load_bytes", new=raise_io):
-        cfg = {"tls_root_certs": "/bad/path"}
-        with pytest.raises(IOError):
-            await get_session_parameters(cfg)
-
-
-@pytest.mark.asyncio
-async def test_get_session_parameters_auth_token_from_env_var(
-    session_manager, monkeypatch
-):
-    """Test auth_token is sourced from environment variable when auth_token_env_var is set."""
-    env_var_name = "MY_TEST_TOKEN_VAR"
-    expected_token = "token_from_environment"
-    monkeypatch.setenv(env_var_name, expected_token)
-
-    worker_cfg = {
-        "auth_token_env_var": env_var_name,
-        # As per config validation, auth_token should not be present if auth_token_env_var is.
-    }
-    params = await get_session_parameters(worker_cfg)
-    assert params["auth_token"] == expected_token
-    monkeypatch.delenv(env_var_name)  # Clean up
-
-
-@pytest.mark.asyncio
-async def test_get_session_parameters_auth_token_env_var_not_set(
-    session_manager, monkeypatch, caplog
-):
-    """Test auth_token is empty and warning logged if auth_token_env_var is set but env var is not."""
-    env_var_name = "MY_MISSING_TOKEN_VAR"
-    monkeypatch.delenv(env_var_name, raising=False)  # Ensure it's not set
-
-    worker_cfg = {
-        "auth_token_env_var": env_var_name,
-    }
-    params = await get_session_parameters(worker_cfg)
-    assert params["auth_token"] == ""
-    assert (
-        f"Environment variable {env_var_name} specified for auth_token but not found. Using empty token."
-        in caplog.text
-    )
-
-
-@pytest.mark.asyncio
-async def test_get_session_parameters_auth_token_from_config(session_manager):
-    """Test auth_token is sourced from config when auth_token_env_var is not set."""
-    expected_token = "token_from_config_direct"
-    worker_cfg = {
-        "auth_token": expected_token,
-    }
-    params = await get_session_parameters(worker_cfg)
-    assert params["auth_token"] == expected_token
-
-
-@pytest.mark.asyncio
-async def test_get_session_parameters_no_auth_token_provided(session_manager):
-    """Test auth_token is empty if neither auth_token nor auth_token_env_var is provided."""
-    worker_cfg = {"host": "localhost"}  # Some other config, but no auth token fields
-    params = await get_session_parameters(worker_cfg)
-    assert params["auth_token"] == ""
 
 
 @pytest.mark.asyncio
 async def test_create_session_error():
     # Patch Session to raise
     with patch(
-        "deephaven_mcp.sessions._lifecycle_community.Session",
+        "deephaven_mcp.sessions._sessions.Session",
         new=MagicMock(side_effect=RuntimeError("fail")),
     ):
         with pytest.raises(SessionCreationError) as exc_info:
             await create_session(host="localhost")
         # Check error message contains context and original error
-        assert "Failed to create Deephaven Session" in str(exc_info.value)
+        assert "Failed to create Deephaven Community (Core) Session" in str(exc_info.value)
         assert "fail" in str(exc_info.value) or (exc_info.value.__cause__ and "fail" in str(exc_info.value.__cause__))
 
 
@@ -295,7 +184,7 @@ async def test_get_or_create_session_liveness_exception(session_manager, caplog)
     session_manager._config_manager.get_config = AsyncMock(
         return_value={"community_sessions": {"foo": {"host": "localhost"}}}
     )
-    with patch("deephaven_mcp.sessions._lifecycle_community.Session", new=MagicMock()):
+    with patch("deephaven_mcp.sessions._lifecycle.community.Session", new=MagicMock()):
         await session_manager.get_or_create_session("foo")
         assert any(
             "Error checking session liveness" in r for r in caplog.text.splitlines()
@@ -335,11 +224,11 @@ async def test_get_or_create_session_creates_new(session_manager):
     mock_session = MagicMock()
     with (
         patch(
-            "deephaven_mcp.sessions._lifecycle_community.get_session_parameters",
+            "deephaven_mcp.sessions._lifecycle.community._get_session_parameters",
             new=AsyncMock(return_value={"host": "localhost"}),
         ),
         patch(
-            "deephaven_mcp.sessions._lifecycle_community.Session",
+            "deephaven_mcp.sessions._lifecycle.community.Session",
             new=MagicMock(return_value=mock_session),
         ),
     ):
@@ -363,11 +252,11 @@ async def test_get_or_create_session_handles_dead(session_manager):
     mock_session = MagicMock()
     with (
         patch(
-            "deephaven_mcp.sessions._lifecycle_community.get_session_parameters",
+            "deephaven_mcp.sessions._lifecycle.community._get_session_parameters",
             new=AsyncMock(return_value={"host": "localhost"}),
         ),
         patch(
-            "deephaven_mcp.sessions._lifecycle_community.Session",
+            "deephaven_mcp.sessions._lifecycle.community.Session",
             new=MagicMock(return_value=mock_session),
         ),
     ):
